+++ maia/algo/part/test/test_extract_boundary.py
@@ -149,20 +149,22 @@ def test_extract_surf_from_bc(comm):
     part_zones = [parse_yaml_cgns.to_node(part_0)]
     bc_pl = np.array([15,16,9,10])
     expt_face_lngn = [3,5,1,2]
+    expt_face_parent = [25, 27, 13, 14]
     expt_vtx_lngn = [1,2,3,4,5,6,7,8,9,10]
   elif comm.Get_rank() == 1:
     part_zones = [parse_yaml_cgns.to_node(part_1)]
     bc_pl = np.array([15,16])
     expt_face_lngn = [4,6]
+    expt_face_parent = [26, 28]
     expt_vtx_lngn = [11,12,13,6,7,8]
 
   bc_predicate = lambda n: PT.get_value(PT.get_child_from_name(n, 'FamilyName')) == 'WALL'
 
-  bc_face_vtx, bc_face_vtx_idx, bc_face_lngn, bc_coords, bc_vtx_lngn = \
+  bc_face_vtx, bc_face_vtx_idx, bc_face_lngn, bc_face_parent, bc_coords, bc_vtx_lngn = \
   EXB.extract_surf_from_bc(part_zones, bc_predicate, comm)
   
 
-  assert len(bc_face_vtx) == len(bc_face_vtx_idx) == len(bc_face_lngn) == len(bc_coords) == len(bc_vtx_lngn) == 1
+  assert len(bc_face_vtx) == len(bc_face_vtx_idx) == len(bc_face_lngn) == len(bc_face_parent) == len(bc_coords) == len(bc_vtx_lngn) == 1
 
   cx, cy, cz, expt_bc_face_vtx_idx, expt_bc_face_vtx, _ = EXB.extract_faces_mesh(part_zones[0], bc_pl)
   assert (bc_face_vtx_idx[0] == expt_bc_face_vtx_idx).all()
@@ -170,3 +172,4 @@ def test_extract_surf_from_bc(comm):
   assert (bc_coords[0] == np.array([cx,cy,cz]).reshape(-1, order='F')).all()
   assert (bc_face_lngn[0] == expt_face_lngn).all()
   assert (bc_vtx_lngn[0] == expt_vtx_lngn).all()
+  assert (bc_face_parent[0] == expt_face_parent).all()